requirements.txt
# DNS Security Auditor - Requirements
# Python 3.8+

# Core Dependencies
streamlit>=1.28.0
dnspython>=2.4.0
pandas>=2.0.0

# Optional: Enhanced Features
plotly>=5.17.0          # For interactive charts
python-dateutil>=2.8.2  # For date parsing
requests>=2.31.0        # For HTTP requests (if needed)

# Development Dependencies (optional)
pytest>=7.4.0           # For testing
black>=23.0.0           # Code formatting
flake8>=6.0.0           # Linting
mypy>=1.5.0             # Type checking
